+++ src/node/ext/fs/events.py
@@ -1,8 +1,0 @@
-from node.ext.fs.interfaces import IFileAddedEvent
-from zope.interface import implementer
-from zope.lifecycleevent import ObjectAddedEvent
-
-
-@implementer(IFileAddedEvent)
-class FileAddedEvent(ObjectAddedEvent):
-    """Event which gets triggered when file is added to directory."""